unreserve ERROR; flip C/Java/C#/Go back to unsigned kinds + 0xFFFF sentinel
ERROR is no longer a reserved token-kind name — a grammar can declare
a token called ERROR if it likes. EOF stays reserved (kind id 0).

Token kinds in C/Java/C#/Go are unsigned again (uint16_t / int /
ushort / uint16; Java has no unsigned 16-bit primitive so it uses int
in the 0..0xFFFF range). The lex-failure sentinel is the in-band
value 0xFFFF, exposed as a named runtime constant in each backend
(PARSUNA_LEX_ERROR / Lexer.ERROR_KIND / Lexer.ErrorKind /
parsunart.ErrorKind) and lives outside the per-grammar TokenKind
enum so user grammars can't shadow it.

EOF is hardcoded to 0 throughout: dropped the eofKind/EofKind
constructor parameters and the PARSUNA_EOF_KIND macro from every
runtime; generated code no longer threads it through.

C runtime cleanup: Token.kind back to uint16_t, set_contains/
matches_first/recover_to/try_consume signatures take uint16_t,
SENTINEL = 0xFFFF (same value as PARSUNA_LEX_ERROR — distinct
contexts, never compared together).

Docs updated across event_model, grammar_language, usage,
pipeline/{parse,analyze,lower,codegen}.

Bootstrap regenerates byte-identically; 170 Rust tests pass; JSON
example regen succeeds for all four sentinel backends.

+++ runtimes/java/src/main/java/dev/parsuna/runtime/Token.java
@@ -3,17 +3,16 @@ package dev.parsuna.runtime;
 /**
  * A lexed token: kind id, span, and the matched text.
  *
- * <p>{@code kind} is a {@code short} matching the generated TokenKind
- * enum's {@code id} field. The runtime reserves {@code -1} as the "no
- * DFA pattern matched" sentinel that the lexer emits when it can't
- * classify a codepoint; real token kinds (including EOF = 0) are
- * non-negative.
+ * <p>{@code kind} is an {@code int} (treated as unsigned 16-bit) matching
+ * the generated TokenKind enum's {@code id} field. EOF = 0; grammar
+ * tokens have ids starting at 1; the lexer emits {@link Lexer#ERROR_KIND}
+ * (0xFFFF) when no DFA pattern matched at the current position.
  */
 public final class Token {
-    public final short kind;
+    public final int kind;
     public final Span span;
     public final String text;
-    public Token(short kind, Span span, String text) {
+    public Token(int kind, Span span, String text) {
         this.kind = kind;
         this.span = span;
         this.text = text;
